requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
pydantic==2.5.2 

webpage interface http://localhost:8001/chat


to run the program:
1.run the main.py python file
2.make sure ur api key is set for gemini 2.0 flash
